requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
pinecone-client>=3.0.0
